+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # `rethinkdb-job-queue` Change log
 
-## v0.0.2 /
+## v0.0.3 / 2016-08
+
+*   Removed 'master' Queue options, only using 'masterInterval'
+*   masterInterval default options changed from 300 sec to 310
+
+## v0.0.2 / 2016-08-01
 
 *   Removed data from Queue.createJob()
 *   Added the Job.setPayload() method